+++ sth-core/src/sth/SurveyEntry.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class SurveyEntry implements Serializable {
   private static final long serialVersionUID = 201810051538L;
 
-  private float _spentHours;
+  private int _spentHours;
   private String _comment;
 
-  SurveyEntry(float spentHours, String comment) {
+  SurveyEntry(int spentHours, String comment) {
     _spentHours = spentHours;
     _comment = comment;
   }
@@ -16,7 +16,7 @@ public class SurveyEntry implements Serializable {
   /**
    * @return the spent hours
    */
-  public double getSpentHours() {
+  public int getSpentHours() {
     return _spentHours;
   }
